% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shapes.R
\name{get_shapes}
\alias{get_shapes}
\title{Retrieve data from the webtool Swiss Maps Generator}
\usage{
get_shapes(
  shapes = c("country", "cantons", "municipalities", "lakes"),
  projection = "wgs84",
  year = 2020,
  simplify = 0
)
}
\arguments{
\item{shapes}{name of shapes to download. Available are \code{"country"}, \code{"cantons"},
\code{"municipalities"} and \code{"lakes"}.}

\item{projection}{name of projection. Available are \code{"wgs84"} and \code{"cartesian"}.}

\item{year}{year of geodata. Available from 2010.}

\item{simplify}{degree of simplification of shapes. Ranging from \code{0} (no simplification)
to \code{100} (maximal degree of simplification).}
}
\value{
If only one type of shapes is requested, a \code{Simple feature collection} is returned.
    For multiple shapes, a list of the corresponding collections is returned.
}
\description{
\code{get_shapes} retrieves geodata via the webtool Swiss Maps Generator.
}
\examples{
# Retrieve geodata
get_shapes()

}
